requirement.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==3.0.4
contextvars==2.4
Django==3.2.6
djangorestframework==3.12.4
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.15
pytz==2021.1
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
